notifications: move SaveConfig, GetConfigOrDefault to config.go
